# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Schedule:
    properties:
      end_time:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      start_time:
        type: string
      subject:
        type: string
    type: object
  models.Student:
    properties:
      birth_date:
        type: string
      course_year:
        type: integer
      gender:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      major:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /schedules:
    get:
      consumes:
      - application/json
      description: Get schedules, optionally filtered by group ID
      parameters:
      - description: Filter by Group ID
        in: query
        name: group_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get schedules
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Add a new class schedule entry
      parameters:
      - description: Schedule Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create schedule
      tags:
      - Schedules
  /schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a schedule entry from the database
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete schedule
      tags:
      - Schedules
    patch:
      consumes:
      - application/json
      description: Update an existing schedule entry
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Schedule Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update schedule
      tags:
      - Schedules
  /students:
    get:
      consumes:
      - application/json
      description: Retrieve students with optional filtering by group, major, and
        course year
      parameters:
      - description: Filter by Group ID
        in: query
        name: group_id
        type: integer
      - description: Filter by Major
        in: query
        name: major
        type: string
      - description: Filter by Course Year
        in: query
        name: course_year
        type: integer
      - description: Limit number of results (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Add a new student record to the database
      parameters:
      - description: Student Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a student
      tags:
      - Students
  /students/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a student record from the database
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a student
      tags:
      - Students
    get:
      consumes:
      - application/json
      description: Get details of a student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a student
      tags:
      - Students
    patch:
      consumes:
      - application/json
      description: Update details of an existing student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Student Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a student
      tags:
      - Students
  /students/{id}/gpa:
    get:
      consumes:
      - application/json
      description: Calculate and return the GPA for a specific student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Student GPA
      tags:
      - Students
...
swagger: "2.0"